% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddtCreateDirs.R
\name{ddtCreateDirs}
\alias{ddtCreateDirs}
\title{Initialize directories for DDT}
\usage{
ddtCreateDirs(ddtPath, testCaseName = "testCase1")
}
\arguments{
\item{ddtPath}{path where tests inputs and outputs should be stored}

\item{testCaseName}{name of test case for DDT}
}
\description{
Initialize directories for DDT
}
\examples{
ddtCreateDirs("./tests/ddt")
}
\seealso{
\code{\link{ddtRun}}
}
