% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReadSaveObj.R
\name{getReadSaveObj.Default}
\alias{getReadSaveObj.Default}
\title{Run singple test under given location}
\usage{
getReadSaveObj.Default()
}
\description{
Run singple test under given location
}
\examples{
ddtRunSingle(".", function(){}, function(){}, getReadSaveObj.CSV ) # run single test
}
\seealso{
\code{\link{ddtRun}}
}
