% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReadSaveObj.JSON.R
\name{getReadSaveObj.JSON}
\alias{getReadSaveObj.JSON}
\title{Reads/Writes objects in JSON format}
\usage{
getReadSaveObj.JSON()
}
\description{
Reads/Writes objects in JSON format
}
\examples{
ddtRunSingle(".", function(){}, function(){}, getReadSaveObj.CSV ) # run single test
}
\seealso{
\code{\link{ddtRun}}
}
