% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReadSaveObj.CSV.R, R/getReadSaveObj.JSON.R,
%   R/getReadSaveObj.R, R/getReadSaveObj.RDS.R
\name{getReadSaveObj.CSV}
\alias{getReadSaveObj.CSV}
\alias{getReadSaveObj.JSON}
\alias{getReadSaveObj}
\alias{getReadSaveObj.RDS}
\title{Reads/Writes objects in JSON format}
\usage{
getReadSaveObj.CSV()

getReadSaveObj.JSON()

getReadSaveObj()

getReadSaveObj.RDS()
}
\description{
Reads/Writes objects in JSON format
}
\section{Functions}{
\itemize{
\item \code{getReadSaveObj.CSV}: reading/writing from/to simple CSV file

\item \code{getReadSaveObj.JSON}: JSON serializer

\item \code{getReadSaveObj.RDS}: RDS serializer
}}

\examples{
#ddtRunSingle(".", function(){}, function(){}, getReadSaveObj ) # run single test
}
\seealso{
\code{\link{ddtRun}}
}
