% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddtRun.R
\name{ddtRun}
\alias{ddtRun}
\title{Run tests under given location}
\usage{
ddtRun(ddtsDir, preload, compare, expectedFileName = "actual.csv",
  readSaveObj = NULL)
}
\arguments{
\item{ddtsDir}{(characters) path where tests inputs and outputs should be stored}

\item{preload}{(function) loading function}

\item{compare}{(function) compared of actual and expected objects}

\item{expectedFileName}{(character) generated file}

\item{readSaveObj}{(list of functions )}
}
\description{
Run tests under given location
}
\examples{
ddtsDir <- './test/ddt'  #location where DDT are
x <- 1:20
y <- x*2 + rnorm(20)
preload <- function(){ return(lm(y~x)) }
compare <- function(expected, actual){
 if(all(abs(expected$coefficients - actual$coefficients) < 0.001) ){
   print("Same")
 } else {
   stop("wrong coefficents. Expected ", print(expected$coefficients) , " but is ",  print(actual$coefficients))
 }}
ddtRun(ddtsDir, preload, compare) # run all test
}
\seealso{
\code{\link{ddtRunSingle}} \code{\link{getReadSaveObj.Default}}

Other ddtRun: \code{\link{ddtRunSingle}}
}
