% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddtRunSingle.R
\name{ddtRunSingle}
\alias{ddtRunSingle}
\title{Run singple test under given location}
\usage{
ddtRunSingle(ddtDir, preload, compare, expectedFileName, readSaveObj = NULL)
}
\arguments{
\item{preload}{(function) loading function}

\item{compare}{(function) compared of actual and expected objects}

\item{expectedFileName}{(character) generated file}

\item{readSaveObj}{(list of functions ) please look at \code{\link{getReadSaveObj.default}}}

\item{ddtsDir}{(characters) path where tests inputs and outputs should be stored}
}
\description{
Run singple test under given location
}
\examples{
ddtsDir <- './test/ddt/testCase1'  #location where single DDT is
reload <- function(){ return(lm(y~x, data=read.csv("example_input.csv")))  }
compare <- function(expected, actual){
 if(all(abs(expected$coefficients - actual$coefficients) < 0.001) ){
   print("Same")
 } else {
   stop("wrong coefficents. Expected ", print(expected$coefficients) , " but is ",  print(actual$coefficients))
 }}
ddtRunSingle(ddtsDir, preload, compare) # run single test
}
\seealso{
\code{\link{ddtRun}}

Other ddtRun: \code{\link{ddtRun}}
}
