% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddtRunSingle.R
\name{ddtRunSingle}
\alias{ddtRunSingle}
\title{Run single test under given location}
\usage{
ddtRunSingle(ddtDir, preload, compare, expectedFileName = "data.txt",
  readSaveObj = NULL)
}
\arguments{
\item{ddtDir}{(characters) path where tests inputs and outputs should be stored}

\item{preload}{(function) loading function}

\item{compare}{(function) compared of actual and expected objects}

\item{expectedFileName}{(character) generated file}

\item{readSaveObj}{(list of functions )}
}
\description{
Run single test under given location
}
\examples{
ddtDir <- './tests/example_in_doc/testCase_1'  #location where single DDT is
x <- 1:20
y <- x*2 + rnorm(20)
preload <- function(){ return(lm(y~x)) }
compare <- function(expected, actual){
 if(all(abs(expected$coefficients - actual$coefficients) < 0.001) ){
   print("Same")
 } else {
   stop("wrong coefficents. Expected ", expected$coefficients, " but is ",  actual$coefficients)
 }}
# ddtRunSingle(ddtDir, preload, compare) # will create required files
}
\seealso{
\code{\link{ddtRun}} \code{\link{getReadSaveObj}}

Other ddtRun: \code{\link{ddtRun}}
}
